Implemented Raw Data Export for Multiple Cells (#204)
* Added resample parameter

Added resample parameter to API GET request for cell data retrieval. This update includes backend support for handling the resample option, allowing the possibility for users to specify a resampling strategy. When resample=none is specified, the server now returns raw data.

* Added validation & default value

* Fixed getCellData for export csv

added toHTTP to startTime & endTime to ensure these are sent in the right format

* Formatted code